Improved UI, Sessions, Async map creation

+++ src/main/java/fr/curie/navicell/storage/FileSystemStorageService.java
@@ -3,6 +3,7 @@ package fr.curie.navicell.storage;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.FileOutputStream;
 import java.net.MalformedURLException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -23,15 +24,17 @@ import org.springframework.web.multipart.MultipartFile;
 public class FileSystemStorageService implements StorageService {
 
 	private final Path rootLocation;
+	private final Path dataLocation;
 
 	@Autowired
 	public FileSystemStorageService(StorageProperties properties) {
-		System.out.println("Root location : " + properties.getLocation());
-		this.rootLocation = Paths.get(properties.getLocation());
+		System.out.println("Root location : " + properties.getMapsLocation());
+		this.rootLocation = Paths.get(properties.getMapsLocation());
+		this.dataLocation = Paths.get(properties.getDataLocation());
 	}
 
 	@Override
-	public Path store(MultipartFile file, String folder, String name) {
+	public Path storeMapFile(MultipartFile file, String folder, String name) {
 		
 		String filename;
 		if (name == null) { 
@@ -60,9 +63,40 @@ public class FileSystemStorageService implements StorageService {
 			throw new StorageException("Failed to store file " + filename, e);
 		}
 	}
+	
+	@Override
+	public Path storeMapFile(byte[] file, String folder, String name) {
+		
+		String filename = name;
+		
+		try {
+			if (file.length == 0) {
+				throw new StorageException("Failed to store empty file " + filename);
+			}
+			if (filename.contains("..")) {
+				// This is a security check
+				throw new StorageException(
+						"Cannot store file with relative path outside current directory "
+								+ filename);
+			}
+			File outputFile = new File(this.rootLocation.resolve(folder + File.separatorChar + filename).toString());
+			try (FileOutputStream outputStream = new FileOutputStream(outputFile)) {
+				outputStream.write(file);
+				return this.rootLocation.resolve(folder + File.separatorChar + filename);
+			}
+			// try (InputStream inputStream = file.getInputStream()) {
+			// 	Files.copy(inputStream, this.rootLocation.resolve(folder + File.separatorChar + filename),
+            //         StandardCopyOption.REPLACE_EXISTING);
+                
+			// }
+		}
+		catch (IOException e) {
+			throw new StorageException("Failed to store file " + filename, e);
+		}
+	}
 
 	@Override
-	public Path store(File file, String folder, String name) {
+	public Path storeMapFile(File file, String folder, String name) {
 		
 		String filename;
 		if (name == null) { 
@@ -93,7 +127,40 @@ public class FileSystemStorageService implements StorageService {
 	}
 	
 	@Override
-	public void createFolder(String folder) {
+	public Path storeDataFile(MultipartFile file, String folder, String name) {
+		
+		String filename;
+		if (name == null) { 
+			filename = StringUtils.cleanPath(file.getOriginalFilename()); 
+		} else {
+			filename = name;
+		}
+		
+		try {
+			if (file.isEmpty()) {
+				throw new StorageException("Failed to store empty file " + filename);
+			}
+			if (filename.contains("..")) {
+				// This is a security check
+				throw new StorageException(
+						"Cannot store file with relative path outside current directory "
+								+ filename);
+			}
+			try (InputStream inputStream = file.getInputStream()) {
+				Files.copy(inputStream, this.dataLocation.resolve(folder + File.separatorChar + filename),
+                    StandardCopyOption.REPLACE_EXISTING);
+                return this.dataLocation.resolve(folder + File.separatorChar + filename);
+			}
+		}
+		catch (IOException e) {
+			throw new StorageException("Failed to store file " + filename, e);
+		}
+	}
+
+	
+	
+	@Override
+	public void createMapFolder(String folder) {
 		Path new_path = Paths.get(this.rootLocation.toString(), folder);
 		if (!Files.exists(new_path)) {
 			try {
@@ -107,41 +174,55 @@ public class FileSystemStorageService implements StorageService {
 	}
 	
 	@Override
-	public Stream<Path> loadAll() {
-		try {
-			return Files.walk(this.rootLocation, 1)
-				.filter(path -> !path.equals(this.rootLocation))
-				.map(this.rootLocation::relativize);
+	public void createDataFolder(String folder) {
+		Path new_path = Paths.get(this.dataLocation.toString(), folder);
+		if (!Files.exists(new_path)) {
+			try {
+				Files.createDirectories(new_path);
+			}
+			catch (IOException e) {
+				throw new StorageException("Failed to create directory", e);
+			}
 		}
-		catch (IOException e) {
-			throw new StorageException("Failed to read stored files", e);
-		}
-
+		
 	}
+	
+	// @Override
+	// public Stream<Path> loadAll() {
+	// 	try {
+	// 		return Files.walk(this.rootLocation, 1)
+	// 			.filter(path -> !path.equals(this.rootLocation))
+	// 			.map(this.rootLocation::relativize);
+	// 	}
+	// 	catch (IOException e) {
+	// 		throw new StorageException("Failed to read stored files", e);
+	// 	}
+
+	// }
     
-	@Override
-	public Path load(String filename) {
-		return rootLocation.resolve(filename);
-	}
+	// @Override
+	// public Path loadMap(String filename) {
+	// 	return rootLocation.resolve(filename);
+	// }
 
-	@Override
-	public Resource loadAsResource(String filename) {
-		try {
-			Path file = load(filename);
-			Resource resource = new UrlResource(file.toUri());
-			if (resource.exists() || resource.isReadable()) {
-				return resource;
-			}
-			else {
-				throw new StorageFileNotFoundException(
-						"Could not read file: " + filename);
+	// @Override
+	// public Resource loadAsResource(String filename) {
+	// 	try {
+	// 		Path file = load(filename);
+	// 		Resource resource = new UrlResource(file.toUri());
+	// 		if (resource.exists() || resource.isReadable()) {
+	// 			return resource;
+	// 		}
+	// 		else {
+	// 			throw new StorageFileNotFoundException(
+	// 					"Could not read file: " + filename);
 
-			}
-		}
-		catch (MalformedURLException e) {
-			throw new StorageFileNotFoundException("Could not read file: " + filename, e);
-		}
-	}
+	// 		}
+	// 	}
+	// 	catch (MalformedURLException e) {
+	// 		throw new StorageFileNotFoundException("Could not read file: " + filename, e);
+	// 	}
+	// }
 
 	@Override
 	public void deleteAll() {
@@ -159,17 +240,39 @@ public class FileSystemStorageService implements StorageService {
 	}
 	
 	@Override
-	public Path getLocation() {
+	public Path getMapsLocation() {
 		return rootLocation;
 	}
 	
 	@Override
-	public void deleteByFolder(String folder) {
-		try{
-			FileSystemUtils.deleteRecursively(Paths.get(rootLocation.toString(), folder));
+	public Path getDataLocation() {
+		return dataLocation;
+	}
+	
+	@Override
+	public void deleteMapByFolder(String folder) {
+		if (folder != null)
+		{
+			try{
+				FileSystemUtils.deleteRecursively(Paths.get(rootLocation.toString(), folder));
+			}
+			catch (IOException e) {
+				throw new StorageException("Could not remove folder " + folder);
+			}
 		}
-		catch (IOException e) {
-			throw new StorageException("Could not remove folder " + folder);
+	}
+	
+	@Override
+	public void deleteDataByFolder(String folder) {
+		
+		if (folder != null) {
+			try{
+				
+				FileSystemUtils.deleteRecursively(Paths.get(dataLocation.toString(), folder));
+			}
+			catch (IOException e) {
+				throw new StorageException("Could not remove folder " + folder);
+			}
 		}
 	}
 }
